NNUE training: int16 feature storage + per-file train/val split
Store decoded HalfKAv2_hm feature indices as int16 instead of int64.
Indices span [0, 22527] with -1 padding, so int16 holds them exactly at
1/4 the RAM, letting the full combined multi-generation dataset fit in
memory without subsampling (int64 caused the 24GB OOM once selfplay-base
was added). Legacy int64 feature caches are cast on load; model.forward
casts indices to long() after the padding remap (bit-identical to the
int64 forward, verified).

Split train/val per file (tail cut per dataset, then concatenate) instead
of tailing the concatenation. The old split put 100% of the validation set
in the last file, biasing best-epoch selection toward the newest generation;
the new split makes validation a representative mix of all generations.

Raise --max-records to a 120M safety cap (int16 makes proportional
subsampling unnecessary for the generational pipeline).

+++ tools/training/nnue/model.py
@@ -42,9 +42,12 @@ class NoaNnue(nn.Module):
             self.ft.weight[INPUT_SIZE].zero_()
 
     def forward(self, stm_feats, opp_feats):
-        # -1 padding -> the zero row.
-        stm_feats = torch.where(stm_feats < 0, torch.full_like(stm_feats, INPUT_SIZE), stm_feats)
-        opp_feats = torch.where(opp_feats < 0, torch.full_like(opp_feats, INPUT_SIZE), opp_feats)
+        # -1 padding -> the zero row. Features are stored int16 in host RAM to
+        # save memory; INPUT_SIZE (22528) fits int16, and EmbeddingBag needs
+        # Long indices, so cast after remapping. .long() is a no-op when the
+        # caller already passes int64 (e.g. validate_nnue's on-the-fly batches).
+        stm_feats = torch.where(stm_feats < 0, torch.full_like(stm_feats, INPUT_SIZE), stm_feats).long()
+        opp_feats = torch.where(opp_feats < 0, torch.full_like(opp_feats, INPUT_SIZE), opp_feats).long()
 
         stm = torch.clamp(self.ft(stm_feats) + self.ft_bias, 0.0, 1.0)
         opp = torch.clamp(self.ft(opp_feats) + self.ft_bias, 0.0, 1.0)
